+++ src/main/java/com/runningmate/server/domain/politicians/service/NationalAssemblyService.java
@@ -1,12 +1,8 @@
 package com.runningmate.server.domain.politicians.service;
 
+import com.runningmate.server.domain.politicians.client.NationalAssemblyApiClient;
 import com.runningmate.server.domain.politicians.dto.external.nationassembly.NaRow;
 import com.runningmate.server.domain.politicians.model.Politician;
-import com.runningmate.server.domain.politicians.model.PoliticianDetail;
-import com.runningmate.server.domain.politicians.repository.PoliticianDetailRepository;
-import com.runningmate.server.domain.politicians.repository.PoliticianRepository;
-import com.runningmate.server.domain.politicians.utils.DateUtil;
-import com.runningmate.server.domain.politicians.utils.NaUtil;
 import com.runningmate.server.domain.politicians.utils.PoliticianUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,13 +20,13 @@ import static com.runningmate.server.domain.politicians.utils.DateUtil.*;
 @RequiredArgsConstructor
 public class NationalAssemblyService {
 
-    private final NaUtil naUtil;
+    private final NationalAssemblyApiClient nationalAssemblyApiClient;
     private final PoliticianUtil politicianUtil;
 
     public void saveMemberImg(){
 
         // 1. 국회 api를 통해서 정보를 가져온다.
-        List<NaRow> naRows = naUtil.fetchNaInfo();
+        List<NaRow> naRows = nationalAssemblyApiClient.fetchNaInfo();
         log.info("naRow size {}", naRows.size());
 
         // 2. 기존 DB에서 정치인 정보 가져오기
